requirements.txt
# Web framework and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==1.10.13

# PDF and document processing
PyPDF4==1.27.0
python-docx==1.0.1
pdfplumber==0.10.3
spacy==3.5.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

# System and file operations
psutil==5.9.6
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
tenacity==8.2.3
pandas==2.1.3
zipfile36==0.1.3
python-dotenv==1.0.0
requests==2.31.0
typing-extensions==4.8.0

# AI and OpenAI
openai==1.3.5
ollama==0.3.3

# Path handling
pathlib==1.0.1

# Development tools (consider moving these to a separate dev-requirements.txt)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0